requirements.txt
# PDC Inference Server — Python dependencies
# Phases 1–4

# --- Inference runtime ---
# Install with CUDA support (required for RTX 4060 Ti):
#   Windows (PowerShell):
#     $env:CMAKE_ARGS="-DGGML_CUDA=on"; pip install llama-cpp-python --force-reinstall --no-cache-dir
#   Linux/macOS:
#     CMAKE_ARGS="-DGGML_CUDA=on" pip install llama-cpp-python --force-reinstall --no-cache-dir
llama-cpp-python>=0.2.90

# --- API server ---
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0

# --- Benchmarking client ---
aiohttp>=3.9.0

# --- Speculative decoding ---
numpy>=1.26.0

# --- System monitoring ---
psutil>=5.9.0
pynvml>=11.5.0          # NVIDIA GPU memory stats (requires NVIDIA driver)

# --- Plotting (Phase 5) ---
matplotlib>=3.9.0
pandas>=2.2.0

# --- Dataset download (Phase 4/5) ---
datasets>=2.19.0
huggingface-hub>=0.22.0
